add: extensions folder with StringExtensions
