+++ RandUsers/RandUsers/Data/Mappers/UserMapper.swift
@@ -35,12 +35,3 @@ extension Results {
         }
     }
 }
-
-extension String {
-    func toDate() -> Date? {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        dateFormatter.timeZone = TimeZone(secondsFromGMT: 0)
-        return dateFormatter.date(from: self)
-    }
-}
