+++ RandUsers/RandUsers/Extensions/StringExtensions.swift
@@ -1,0 +1,17 @@
+//
+//  StringExtensions.swift
+//  RandUsers
+//
+//  Created by Guido Fabio on 7/3/25.
+//
+
+import Foundation
+
+extension String {
+    func toDate() -> Date? {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+        dateFormatter.timeZone = TimeZone(secondsFromGMT: 0)
+        return dateFormatter.date(from: self)
+    }
+}
